Add python script for generating k8s config files
There are a number of different configuration options that we are
curious about with respect to running k8s replication controllers.
Currently I can't think of a better way of specifying all of these
options other than creating different config files for all of the
options. However, I definitely don't want to manually generate all of
these config files each time I want to test a new option. So I created a
python script to do it for me.

* It was decided to run this script within a Docker container. The
  generated files will carry over to the host machine because we are
  linking the volume. The use of Docker is done to reduce the number of
  dependencies on any individual host machine seeking to work on this
  thesis.

* We create all of the different configuration files using a templating
  engine. We decided on using `jinja2` as the templating engine. Other
  options include `erb`, but I wanted to use python if possible because
  I imagine I'll be writing the data analysis components with python.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>