Modifications for successful k8s deploy
This commit contains a number of changes made to configuration files,
and configuration scripts, to be able to successfully run our
replication controllers, services, and hpas on k8s. This commit also
includes instructions on how to deploy k8s to EC2. All parts, including
deploying, running replication controllers, running services, and
running hpas with our configuration files are working!

Additionally, this commit contains the creation of a new directory
called containers, which contains Dockerfiles for all of the other
non `test-server` containers I'll want to create during the course of
this thesis. As they are mostly packaging dependencies, they don't have
a ton of code, so that is why they are all packaged in the same
directory instead of separate like `test-server`. This commit contains
the first container, `thesis-python` which contains a number of python
dependencies I'll need for all python code in this thesis, but there
will be more later, particularly when I containerize the traffic
generator so that I can run it on k8s.

Finally, this commit contains updates to the config file templates and
bug fixes to `generate_k8s_config.py`, the configuration generation
script.

tl;dr

* Working configuration files for rcs, svcs, and hpas running
  test-server
* Creation of a new directory called `containers`, and the first
  container in it, `thesis-python`, which is a container including
  helpful python deps.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>