Outline for using jmeter for traffic generation
I've decided that jmeter is the best option for generating the traffic
we need. With the help of the Throughput Shaping Timer plugin, we'll be
able to craft any interesting pattern of traffic that we want.
Additionally, jmeter can be easily Dockerized and run on kubernetes.

This commit contains instructions for creating `test plans` with jmeter
and the Throughput Shaping Timer. Additionally, it contains a
sample-test-plan.jmx file which allows the test plan to be run from the
command line, and requires only slight modifications to create other
traffic patterns.

There is a little bit of work still left to be done re deciding on how
long tests should run and what level of requests should be sent before
we can actually create the final `jmx` files with these patterns, but
once these values are decided, it will be simple to make the test plans.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>