Update traffic request pattern info in eval
Update the evaluation section both with the surface level descriptions
of the traffic request patterns we'll test and a description of how
we'll do it with jmeter. The description of the traffic request patterns
in the independent variables section still has some @TODOs because I
need to finalize the patterns and also some other variables relating to
how long we send traffic for and the largest amount of requests we'll
send.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>