Proposed fix for eru metrics bug
Previously I was calculating eru in `test-server` by trying to find the
user CPU time based the `clock` function in C. But I'm worried that this
value wasn't actually a good indicator, as if there are no other
processes running, then almost all clock ticks will go the `test-server`
application, resulting in a high percentage. This makes it appear that
`eru` is high when it is not.

Instead, we switch to tracking `eru` based on the amount of CPU idle
time. The lower this value, the better the `eru`. We still need to
double check that this idle time value makes sense as an indicator and
is being calculated correctly, yet it seems like a good start.

We also need to update the written portions of the thesis to reflect
the new specific value we track to determine `eru`.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>