Update new prediction method
Previously I'd been writing as if I'd be using the derivative of
previous CPU utilization to make the future prediction. That's not
entirely accurate. It's a similar idea, but instead we're going to model
a linear line of best fit and then predict from that. This commit
updates all sections talking about the future prediction to use the line
of best fit method.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>