Update new configuration for running tests
This commit reflects changes from my first time trying to use jmeter to
run changes on Kubernetes. The positive is that everything is working in
terms of talking to each other. The generated traffic is finding the
pods and the pods are communicating with the database.

A couple of changes were necessary to make that happen.

* I imposed a limit and request of 100M CPU on each pod. With 4 nodes,
  and 1000M on each node, then roughly 40 pods can exist, minus any CPU
  k8s reserves for internal processing (which I think is actually
  relatively substantial).
* I'm using internal DNS to find the `test-server` service, instead of
  using an external HTTP endpoint. The advantage of this is that as long
  as I'm only testing one `test-server` replication controller at a
  time, which I planned on doing anyway, then I don't ever need to edit
  the jmeter test plans to refer to a different endpoint.

An important change still left to make is that I need to vastly decrease
the amount of requests. Currently the peak is at 200 RPS, with a threads
account on 1000, which is way way too many requests. I'll need to fix
this before continuing.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>