Work on processing
This commit contains progress on two parts of the processing. First, it
makes some updates to how we record data to influxdb, to ensure that we
are formatting it correctly, and also that we are recording all of the
data that we need.

Additionally, while not tested, this commit adds the parts to
`evaluate.py` that retrieve all of the data from influxdb and calculate
the sum of eru and qos. This just leaves us the code for creating the
graphs and creating the output summary statistics.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>