Add versioning to trials
This commit adds configuration to `test-server` and `evaluate.py` so
that we can run multiple versions. However, it is important to note that
the `evaluate.py` will not work until an actual versioned trial is run,
so if any evaluations are necessary in the meantime, we will need to do
them with a previous git commit.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>