Update future work and implementation for v2
Version 2 of my trials reduced the forbidden scaling window to 30s, so
include those changes within the written portion of my thesis. In
addition, add a small discussion to the future work about the
possibility of recording TotalCPUUtilization, not AverageCPUUtilization,
with respect to what is used to make predictions about future CPU
Utilization.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>